WindEditor/EditorCore: Added a template for a Bits field which assumes you'll never have a bits field over 8 bits in length (ie: a byte)

+++ EditorCore/WindWaker/Loaders/MapEntityLoader.cs
@@ -234,7 +234,7 @@ namespace WEditor.WindWaker.Loaders
                         break;
 
                     case PropertyType.Bits:
-                        value = (int)reader.ReadBits(templateProperty.Length);
+                        value = (byte)reader.ReadBits(templateProperty.Length);
                         break;
                 }
